+++ src/main/java/com/gkhn/issuemanagement/repository/ProjectRepository.java
@@ -1,15 +1,15 @@
 package com.gkhn.issuemanagement.repository;
 
-import com.gkhn.issuemanagement.entity.Project;
+import java.util.List;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-import java.util.List;
+import com.gkhn.issuemanagement.entity.Project;
 //@Repository Autowired için gerekli. Ancak JPARepository Spring 2.0 dan sonra gerek kalmadı.
-public interface ProjectRepository extends JpaRepository<Project, BigInteger> {
+public interface ProjectRepository extends JpaRepository<Project, Long> {
     List<Project> getAllByProjectCode(String projectCode);
 //    List<Project> getAllByProjectCodeAndIdNotNull(String projectCode);
 //    List<Project> getAllByProjectCodeAnAndProjectNameContains(String projectCode, String name);
